Add a config knob for status checks that must pass
In the default configuration Merge Pal is called and will act on every status
check.

This can burn through the API rate limit if there are some status checks
that you don't care about -- for example, the checks are arranged as a tree
and you only care that the root of the tree has passed (which means all the
others have passed to).

Introduce a configuration option, `passing_status_checks`, which is a list of
0 or more checks that are required. If set, status updates that don't
mention these checks will be ignored.