Reduce API calls by only calling client.pulls.get() when necessary
Previously, `mergeIfReady` received the PR number as a parameter. This
meant that it always had to call `client.pulls.get()` to fetch the PR
information.

Specifically, it needed the `mergeable` and `mergeable_state` fields,
used by the call to `canMerge()`.

However, those fields exist on some payloads given to the action, so this
call was not always needed.

Specifically, if it's a `pull_request` or `pull_request_review` action then
the webhook payload contains those fields for the PR.

But if it's a PR returned by `client.pulls.list()` then it doesn't have
those fields, and `client.pulls.get()` needs to be called.

Fixing this touches almost everything.

1. Change `isEnabledForPR` to take the label names to match against, not
   the full PR. This simplifies the type of data that function deals with,
   and is consistent with an approach where functions only receive the
   specific data they need.

2. Remove the custom types from `types.ts` and use the types from the
   Octokit libraries. This includes the types from `@octokit/rest` which
   are relevant for the values returned using the `client` API, and
   `@octokit/webhooks` which contains types for the data provided to the
   action by GitHub.

3. In `mergeIfReady`, only call `client.pulls.get()` if the provided PR
   is a `PullsListResponseItem` (and therefore is missing the fields).

WIP: Always fetch PR data using the client API

Real world experience shows that the `mergeable*` keys in the webhook payload
can't be trusted.